requirements.txt
Adafruit_DHT
pyyaml
bluepy
smbus2
paho-mqtt
